spotify_api_data_extract/requirements.txt
google-cloud-bigquery
google-cloud-storage
spotipy
functions_framework

